+++ acceptance/python/src/test_divisible_asset.py
@@ -28,6 +28,7 @@ from planetmint_driver.exceptions import BadRequest
 # For this test case we import and use the Python Driver.
 from planetmint_driver import Planetmint
 from planetmint_driver.crypto import generate_keypair
+from ipld import multihash, marshal
 
 
 def test_divisible_assets():
@@ -48,10 +49,14 @@ def test_divisible_assets():
     # the bike for one hour.
 
     bike_token = {
-        "data": {
-            "token_for": {"bike": {"serial_number": 420420}},
-            "description": "Time share token. Each token equals one hour of riding.",
-        },
+        "data": multihash(
+            marshal(
+                {
+                    "token_for": {"bike": {"serial_number": 420420}},
+                    "description": "Time share token. Each token equals one hour of riding.",
+                }
+            )
+        ),
     }
 
     # She prepares a `CREATE` transaction and issues 10 tokens.