+++ docs/root/generate_http_server_api_documentation.py
@@ -15,6 +15,7 @@ from planetmint import lib
 from planetmint.transactions.types.assets.create import Create
 from planetmint.transactions.types.assets.transfer import Transfer
 from planetmint.web import server
+from ipld import multihash, marshal
 
 
 TPLS = {}
@@ -157,10 +158,12 @@ def main():
     ctx["api_index"] = pretty_json(json.loads(res.data.decode()))
 
     # tx create
+    from ipld import marshal, multihash
+
     privkey = "CfdqtD7sS7FgkMoGPXw55MVGGFwQLAoHYTcBhZDtF99Z"
     pubkey = "4K9sWUMFwTgaDGPfdynrbxWqWS6sWmKbZoTjxLtVUibD"
-    asset = {"msg": "Hello Planetmint!"}
-    tx = Create.generate([pubkey], [([pubkey], 1)], asset=asset, metadata={"sequence": 0})
+    asset = {"data": multihash(marshal({"msg": "Hello Planetmint!"}))}
+    tx = Create.generate([pubkey], [([pubkey], 1)], asset=asset, metadata=multihash(marshal({"sequence": 0})))
     tx = tx.sign([privkey])
     ctx["tx"] = pretty_json(tx.to_dict())
     ctx["public_keys"] = tx.outputs[0].public_keys[0]
@@ -176,7 +179,9 @@ def main():
         fulfills=TransactionLink(txid=tx.id, output=cid),
         owners_before=tx.outputs[cid].public_keys,
     )
-    tx_transfer = Transfer.generate([input_], [([pubkey_transfer], 1)], asset_id=tx.id, metadata={"sequence": 1})
+    tx_transfer = Transfer.generate(
+        [input_], [([pubkey_transfer], 1)], asset_id=tx.id, metadata=multihash(marshal({"sequence": 1}))
+    )
     tx_transfer = tx_transfer.sign([privkey])
     ctx["tx_transfer"] = pretty_json(tx_transfer.to_dict())
     ctx["public_keys_transfer"] = tx_transfer.outputs[0].public_keys[0]
@@ -192,7 +197,7 @@ def main():
         owners_before=tx_transfer.outputs[cid].public_keys,
     )
     tx_transfer_last = Transfer.generate(
-        [input_], [([pubkey_transfer_last], 1)], asset_id=tx.id, metadata={"sequence": 2}
+        [input_], [([pubkey_transfer_last], 1)], asset_id=tx.id, metadata=multihash(marshal({"sequence": 2}))
     )
     tx_transfer_last = tx_transfer_last.sign([privkey_transfer])
     ctx["tx_transfer_last"] = pretty_json(tx_transfer_last.to_dict())