+++ tests/assets/test_zenroom_signing.py
@@ -6,6 +6,7 @@ from zenroom import zencode_exec
 from cryptoconditions.types.ed25519 import Ed25519Sha256
 from cryptoconditions.types.zenroom import ZenroomSha256
 from planetmint.transactions.common.crypto import generate_key_pair
+from ipld import multihash, marshal
 
 CONDITION_SCRIPT = """Scenario 'ecdh': create the signature of an object
     Given I have the 'keyring'
@@ -101,8 +102,8 @@ def test_zenroom_signing():
     metadata = {"result": {"output": ["ok"]}}
     token_creation_tx = {
         "operation": "CREATE",
-        "asset": {"data": {"test": "my asset"}},
-        "metadata": metadata,
+        "asset": {"data": multihash(marshal({"test": "my asset"}))},
+        "metadata": multihash(marshal(metadata)),
         "script": script_,
         "outputs": [
             output,
@@ -170,9 +171,6 @@ def test_zenroom_signing():
         planet.validate_transaction(tx_obj)
     except ValidationError as e:
         print("Invalid transaction ({}): {}".format(type(e).__name__, e))
-        assert ()
-    except e:
-        print(f"Exception : {e}")
         assert ()
 
     print(f"VALIDATED : {tx_obj}")