+++ tests/common/conftest.py
@@ -23,9 +23,9 @@ CC_FULFILLMENT_URI = (
 )
 CC_CONDITION_URI = "ni:///sha-256;" "eZI5q6j8T_fqv7xMROaei9_tmTMk4S7WR5Kr4onPHV8" "?fpt=ed25519-sha-256&cost=131072"
 
-ASSET_DEFINITION = {"data": {"definition": "Asset definition"}}
+ASSET_DEFINITION = {"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"}
 
-DATA = {"msg": "Hello Planetmint!"}
+DATA = "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"
 
 
 @pytest.fixture
@@ -189,11 +189,7 @@ def dummy_transaction():
 @pytest.fixture
 def unfulfilled_transaction():
     return {
-        "asset": {
-            "data": {
-                "msg": "Hello Planetmint!",
-            }
-        },
+        "asset": {"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"},
         "id": None,
         "inputs": [
             {
@@ -229,11 +225,7 @@ def unfulfilled_transaction():
 @pytest.fixture
 def fulfilled_transaction():
     return {
-        "asset": {
-            "data": {
-                "msg": "Hello Planetmint!",
-            }
-        },
+        "asset": {"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"},
         "id": None,
         "inputs": [
             {
@@ -292,11 +284,7 @@ def fulfilled_transaction():
 )
 def tri_state_transaction(request):
     tx = {
-        "asset": {
-            "data": {
-                "msg": "Hello Planetmint!",
-            }
-        },
+        "asset": {"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"},
         "id": None,
         "inputs": [
             {"fulfillment": None, "fulfills": None, "owners_before": ["JEAkEJqLbbgDRAtMm8YAjGp759Aq2qTn9eaEHUj2XePE"]}