+++ tests/validation/test_transaction_structure.py
@@ -20,6 +20,7 @@ from unittest.mock import MagicMock
 from planetmint.transactions.common.exceptions import AmountError, SchemaValidationError, ThresholdTooDeep
 from planetmint.transactions.common.transaction import Transaction
 from planetmint.transactions.common.utils import _fulfillment_to_details, _fulfillment_from_details
+from ipld import marshal, multihash
 
 ################################################################################
 # Helper functions
@@ -85,7 +86,7 @@ def test_validate_invalid_operation(b, create_tx, alice):
 
 
 def test_validate_fails_metadata_empty_dict(b, create_tx, alice):
-    create_tx.metadata = {"a": 1}
+    create_tx.metadata = multihash(marshal({"a": 1}))
     signed_tx = create_tx.sign([alice.private_key])
     validate(signed_tx)
 
@@ -129,9 +130,10 @@ def test_create_tx_no_asset_id(b, create_tx, alice):
 
 
 def test_create_tx_asset_type(b, create_tx, alice):
-    create_tx.asset["data"] = "a"
+    create_tx.asset["data"] = multihash(marshal({"a": ""}))
     signed_tx = create_tx.sign([alice.private_key])
-    validate_raises(signed_tx)
+    validate(signed_tx)
+    # validate_raises(signed_tx)
 
 
 def test_create_tx_no_asset_data(b, create_tx, alice):