+++ tests/web/test_assets.py
@@ -5,6 +5,7 @@
 
 import pytest
 from planetmint.transactions.types.assets.create import Create
+from ipld import marshal, multihash
 
 ASSETS_ENDPOINT = "/api/v1/assets/"
 
@@ -29,24 +30,24 @@ def test_get_assets_tendermint(client, b, alice):
     assert res.status_code == 200
 
     # create asset
-    asset = {"msg": "abc"}
+    asset = {"data": multihash(marshal({"msg": "abc"}))}
     tx = Create.generate([alice.public_key], [([alice.public_key], 1)], asset=asset).sign([alice.private_key])
 
     b.store_bulk_transactions([tx])
 
     # test that asset is returned
-    res = client.get(ASSETS_ENDPOINT + "?search=abc")
+    res = client.get(ASSETS_ENDPOINT + "?search=" + asset["data"])
     assert res.status_code == 200
     assert len(res.json) == 1
-    assert res.json[0] == {"data": {"msg": "abc"}, "id": tx.id}
+    assert res.json[0] == {"data": asset["data"], "id": tx.id}
 
 
 @pytest.mark.bdb
 def test_get_assets_limit_tendermint(client, b, alice):
 
     # create two assets
-    asset1 = {"msg": "abc 1"}
-    asset2 = {"msg": "abc 2"}
+    asset1 = {"data": multihash(marshal({"msg": "abc 1"}))}
+    asset2 = {"data": multihash(marshal({"msg": "abc 2"}))}
     tx1 = Create.generate([alice.public_key], [([alice.public_key], 1)], asset=asset1).sign([alice.private_key])
     tx2 = Create.generate([alice.public_key], [([alice.public_key], 1)], asset=asset2).sign([alice.private_key])
 
@@ -54,11 +55,11 @@ def test_get_assets_limit_tendermint(client, b, alice):
     b.store_bulk_transactions([tx2])
 
     # test that both assets are returned without limit
-    res = client.get(ASSETS_ENDPOINT + "?search=abc")
+    res = client.get(ASSETS_ENDPOINT + "?search=" + asset1["data"])
     assert res.status_code == 200
-    assert len(res.json) == 2
+    assert len(res.json) == 1
 
     # test that only one asset is returned when using limit=1
-    res = client.get(ASSETS_ENDPOINT + "?search=abc&limit=1")
+    res = client.get(ASSETS_ENDPOINT + "?search=" + asset1["data"] + "&limit=1")
     assert res.status_code == 200
     assert len(res.json) == 1